requirements.txt
asgiref==3.5.2
boto3==1.26.57
botocore==1.29.57
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.0.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-dbbackup==4.0.1
django-filter==22.1
django-on-heroku==1.1.2
django-phonenumber-field==6.3.0
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
phonenumberslite==8.12.51
psycopg2-binary==2.9.5
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
PyYAML==6.0
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
whitenoise==6.3.0
